requirements.txt
discord.py>=2.0.0
pytz
fredapi
pandas
finvizfinance
aiohttp
